Finish ForClient 1.34.1. Last fork for lastclient offi

+++ src/database/dynamics/data/BankData.java
@@ -72,7 +72,7 @@ public class BankData extends AbstractDAO<Object> {
             result = getData("SELECT * FROM `banks` WHERE id = '" + guid + "'");
             ResultSet RS = result.resultSet;
             if (RS.next()) {
-                get = RS.getInt("kamas") + "@" + RS.getString("items");
+                get = RS.getLong("kamas") + "@" + RS.getString("items");
             }
         } catch (SQLException e) {
             super.sendError("BankData getWaitingAccount", e);
